+++ actions/authorization.go
@@ -1,0 +1,10 @@
+package actions
+
+///////////////////////////
+// AUTHORIZATION_SIGN_IN //
+///////////////////////////
+
+// CreateAuthorizationSignInFailure .
+func CreateAuthorizationSignInFailure(errors []string) *Action {
+	return createFailureAction(AuthorizationSignInFailure, errors)
+}
