+++ actions/group.go
@@ -1,5 +1,23 @@
 package actions
 
+//////////////////////
+// GROUP_ADD_MEMBER //
+//////////////////////
+
+// CreateGroupAddMemberFailure .
+func CreateGroupAddMemberFailure(errors []string) *Action {
+	return createFailureAction(GroupAddMemberFailure, errors)
+}
+
+// CreateGroupAddMemberSuccess .
+func CreateGroupAddMemberSuccess() *Action {
+	return createEmptyAction(GroupAddMemberSuccess)
+}
+
+//////////////////
+// GROUP_CREATE //
+//////////////////
+
 // CreateGroupCreateFailure .
 func CreateGroupCreateFailure(errors []string) *Action {
 	return createFailureAction(GroupCreateFailure, errors)
@@ -24,4 +42,18 @@ func CreateGroupCreateSuccess(
 		Payload: payload,
 		Type:    GroupCreateSuccess,
 	}
+}
+
+/////////////////
+// GROUP_LEAVE //
+/////////////////
+
+// CreateGroupLeaveFailure .
+func CreateGroupLeaveFailure(errors []string) *Action {
+	return createFailureAction(GroupLeaveFailure, errors)
+}
+
+// CreateGroupLeaveSuccess .
+func CreateGroupLeaveSuccess() *Action {
+	return createEmptyAction(GroupLeaveSuccess)
 }
