+++ actions/messaging.go
@@ -1,5 +1,25 @@
 package actions
 
+import "github.com/hooligram/hooligram-server/db"
+
+//////////////////////////////
+// CREATE_MESSAGING_DELIVER //
+//////////////////////////////
+
+// CreateMessagingDeliverRequest .
+func CreateMessagingDeliverRequest(message *db.Message) *Action {
+	payload := make(map[string]interface{})
+	payload["content"] = message.Content
+	payload["date_created"] = message.DateCreated
+	payload["id"] = message.ID
+	payload["sender_id"] = message.SenderID
+
+	return &Action{
+		Payload: payload,
+		Type:    MessagingDeliverRequest,
+	}
+}
+
 // CreateMessagingDeliverSuccessFailure .
 func CreateMessagingDeliverSuccessFailure(errors []string) *Action {
 	return createFailureAction(MessagingDeliverSuccessFailure, errors)
@@ -9,6 +29,10 @@ func CreateMessagingDeliverSuccessFailure(errors []string) *Action {
 func CreateMessagingDeliverSuccessSuccess() *Action {
 	return createEmptyAction(MessagingDeliverSuccessSuccess)
 }
+
+///////////////////////////
+// CREATE_MESSAGING_SEND //
+///////////////////////////
 
 // CreateMessagingSendFailure .
 func CreateMessagingSendFailure(errors []string) *Action {
