+++ actions/verification.go
@@ -1,5 +1,9 @@
 package actions
 
+///////////////////////////////
+// VERIFICATION_REQUEST_CODE //
+///////////////////////////////
+
 // CreateVerificationRequestCodeFailure .
 func CreateVerificationRequestCodeFailure(errors []string) *Action {
 	return createFailureAction(VerificationRequestCodeFailure, errors)
@@ -9,6 +13,10 @@ func CreateVerificationRequestCodeFailure(errors []string) *Action {
 func CreateVerificationRequestCodeSuccess() *Action {
 	return createEmptyAction(VerificationRequestCodeSuccess)
 }
+
+//////////////////////////////
+// VERIFICATION_SUBMIT_CODE //
+//////////////////////////////
 
 // CreateVerificationSubmitCodeFailure .
 func CreateVerificationSubmitCodeFailure(errors []string) *Action {
